fix: Simplify Solana sign-in message generation
- Refocus on required fields 'DOMAIN' and 'ADDRESS' only
- Removed handling for optional fields from message generation to ensure compatibility with Solana SDK's parsing method
- Adjusted the  function accordingly

This change resolves the verification mismatch issue encountered when using the SDK's  function.
