+++ frontend/src/styles/GlobalStyles.css
@@ -1,0 +1,62 @@
+body {
+  font-size: 1rem;
+  font-family: "Inter", sans-serif !important;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column; }
+  body footer {
+    margin-top: auto; }
+  body a {
+    text-decoration: none; }
+  body img {
+    display: block;
+    max-width: 100%;
+    height: auto;
+    margin: auto; }
+
+.half {
+  width: 50%; }
+
+.main-container {
+  max-width: 1440px;
+  margin: 0 auto;
+  padding: 0 2rem; }
+
+.dark-bg {
+  background-color: #393D3F;
+  color: #FFFFFF; }
+
+.light-text {
+  color: #FFFFFF; }
+
+.dark-text {
+  color: #000000; }
+
+.alt-bg {
+  background-color: #62929E; }
+
+.btn-primary {
+  background-color: #000000;
+  border-radius: 0.3125rem;
+  color: #FFFFFF;
+  padding: 0.5556rem 1.1111rem;
+  text-align: center; }
+
+.btn-secondary {
+  background-color: #FFFFFF;
+  border-radius: 0.3125rem;
+  color: #000000;
+  padding: 0.5556rem 1.1111rem;
+  text-align: center; }
+
+.hidden,
+.visible {
+  transition: all 0.5s ease; }
+
+.hidden {
+  opacity: 0;
+  transform: translateY(100px); }
+
+.visible {
+  opacity: 1;
+  transform: translateY(0); }
